Refactor to use proper theme colors for dark mode
Updated landing page to be more interactive for users. Added new theme
and design to support vibrant colors and scheme.

Updated print functionality and drawer open / close fn so that the
confirmation does not block the print functionality. Updated code to
support closing the drawer when the print fn is called.

Updated colors and css for other areas and kept footer across all pages.

Removed un-used files and images