Add firebase and custom analytics
Added firebase to store custom metrics. We did this to track the user
flow within the application. Added firebase to save those records. To
ensure that the records are not overflowing in the system, we might want
to periodically track and remove unwanted data or archive them.

Added hook to track where the user landed based on user ip address. Since
we do not have any network calls, this ip is not accurate and that is ok
since we are just trying to see where users land within the application.

Added custom analytics metrics and peform cleanup

- Added custom analytics to muiv5 button component. This enables to track
all click events made by a user. We can read this custom metrics from the
firestore / firebase store.

- Added context for navigation analytics. This is done so that we do not
mis-use and / or over-use the provided function. When using a hook, the
boilerplate logic got too cumbersome to facilitate and hence we used
context for a more streamlined approach. The application is wrapped within
this context window and everytime the location of the url changes, the
useEffect performs analytics for the selected user.

- Performed cleanup and updated tech debt code.