11: Create dashboard and professional routes
Added ability to customize aspects of the navbar based on clientConfig
params generated from the .env file. This gives us leverage to build our
application in little chunks and safely deploy them when fully tested.

Renamed pro routes to dashboard as we want to offer the ability to view
dashboard data if users can submit json data or pdf data if we can support
it later on.. Till then it should be hidden behind the feature flag.