53 - Added support for BreadCrumbs Component
Added support for BreadCrumbs. This allows the user to navigate quickly
and with ease across the system.

Also added support for FAQ and WhatsNewSection. This two pages although
will not be displayed in the navigation bar, are critical to allow users
to know what is going on with the application.

Updated tech debt as well.