[feature] - 62 - Create Login Routes
- Updated routes to contain a requirements for login in the config section
of each routes.

- Updated helper functions to determine if the user is logged in and display
respective routes / stuffs for the user.

- Created skeleton for properties section of the app.

- Updated .env file in order to support firebase authentication

- Updated routes to create login/signup routes and updated invoicer routes to fall under /invoice

- Created signup and login components.

- Updated home page to display login/create account buttons.

- Updated breadcrumbs

- Added google authentication to the application

- Updated firebase config to support authenticator and analytics db
for invoicer application. We need to find a way to purge the analytics
and it seems that would be much more than what we can handle.

- Update Routes to remove signup and login routes, we didn't need them.
Instead for now, we just support google authentication.

- Updated helper functions to display logged in users. Updated google authentication
functionalities to allow for seamless login / sign out of the application.

- Added small logout button in the AppToolbar.jsx

62 - Added properties portal

- Added basic framiing work for users who are logged into the system. This is
primary for adding a property. I tried to use yup and other validators to
quickly make this form, but it felt like doing it manually would be easiest
and fastest because of the full control of the form change steps.