Updated Edit Property Page
- Added react hook forms for easy form management
- Updated TextFieldWithLabel to support react form hooks
- Updated Tenants and QuickActions to support property rent and
additional charges.
- Added React Hook Forms to Add Property. We should start slowly
integrating this form component in our app since this is a great way
to boost our performance of the app.
- Added Routes for tenant. Also built support for tenants only able to see
their own respective properties and vice versa. Required a lot of
configuration for the ruleset in firebase itself. Further supported ability
to route tenants to the respective destination.
- Added ability to configure routes based on user role. Added some checks to
ensure that a user is redirected to the correct portal ( owner or tenant ) based
on the status of the user role. This is retrieved from the db and hence is
the most accurate information of the user role. However, due to expensive
computational cost, this is done only during the creation of NavBar route elements.
For an easy workflow, all routes are built regardless of whether the user has the
correct route privi or not.

- Added cleanup across various functions and fields for a better and
smoother transition to the workflow.